+++ backend/src/app.py
@@ -1,8 +1,8 @@
 from flask import Flask
 from flask_cors import CORS
 from dotenv import load_dotenv
-from src.controller.colaborador_controller import bp_colaborador
-from src.controller.reembolso_controller import bp_reembolso
+from src.controller.colaborador.colaborador_controller import bp_colaborador
+from src.controller.reembolso.reembolso_controller import bp_reembolso
 from src.model import db
 from config import Config
 from flasgger import Swagger
